重构: 修复了部分距离现象，如声音溢出等问题

+++ src/main/java/net/minecraft/client/resources/sounds/BeeSoundInstance.java
@@ -19,9 +19,9 @@ public abstract class BeeSoundInstance extends AbstractTickableSoundInstance {
     public BeeSoundInstance(final Bee bee, final SoundEvent event, final SoundSource source) {
         super(event, source, SoundInstance.createUnseededRandom());
         this.bee = bee;
-        this.x = (float)bee.getX();
-        this.y = (float)bee.getY();
-        this.z = (float)bee.getZ();
+        this.x = bee.getX();
+        this.y = bee.getY();
+        this.z = bee.getZ();
         this.looping = true;
         this.delay = 0;
         this.volume = 0.0F;
@@ -36,9 +36,9 @@ public abstract class BeeSoundInstance extends AbstractTickableSoundInstance {
         }
 
         if (!this.bee.isRemoved() && !this.hasSwitched) {
-            this.x = (float)this.bee.getX();
-            this.y = (float)this.bee.getY();
-            this.z = (float)this.bee.getZ();
+            this.x = this.bee.getX();
+            this.y = this.bee.getY();
+            this.z = this.bee.getZ();
             float speed = (float)this.bee.getDeltaMovement().horizontalDistance();
             if (speed >= 0.01F) {
                 this.pitch = Mth.lerp(Mth.clamp(speed, this.getMinPitch(), this.getMaxPitch()), this.getMinPitch(), this.getMaxPitch());
